moved context stuff to setupTests
